refactor(twitch): update settings patches
