+++ src/main/kotlin/app/revanced/patches/twitch/misc/settings/resource/patch/SettingsResourcePatch.kt
@@ -8,21 +8,25 @@ import app.revanced.patcher.patch.PatchResult
 import app.revanced.patcher.patch.PatchResultSuccess
 import app.revanced.patcher.patch.ResourcePatch
 import app.revanced.patcher.patch.annotations.DependsOn
+import app.revanced.patches.shared.settings.BasePreference
+import app.revanced.patches.shared.settings.IResource
+import app.revanced.patches.shared.settings.addPreference
+import app.revanced.patches.shared.settings.addResource
+import app.revanced.patches.shared.settings.impl.ArrayResource
+import app.revanced.patches.shared.settings.impl.StringResource
 import app.revanced.patches.twitch.misc.settings.annotations.SettingsCompatibility
-import app.revanced.patches.twitch.misc.settings.components.BasePreference
-import app.revanced.patches.twitch.misc.settings.components.impl.*
 import app.revanced.patches.youtube.misc.manifest.patch.FixLocaleConfigErrorPatch
 import app.revanced.util.resources.ResourceUtils
 import app.revanced.util.resources.ResourceUtils.copyResources
-import org.w3c.dom.Element
 import org.w3c.dom.Node
 
-@Name("twitch-settings-resource-patch")
+@Name("settings-resource-patch")
 @SettingsCompatibility
 @DependsOn([FixLocaleConfigErrorPatch::class])
 @Version("0.0.1")
 class SettingsResourcePatch : ResourcePatch {
     override fun execute(context: ResourceContext): PatchResult {
+
         /* required for self-restart in settings */
         context.xmlEditor["AndroidManifest.xml"].use { editor ->
             editor.file.getElementsByTagName("manifest").item(0).also {
@@ -83,27 +87,8 @@ class SettingsResourcePatch : ResourcePatch {
          *
          * @param arrayResource The array resource to add.
          */
-        fun addArray(arrayResource: ArrayResource) {
-            arraysNode!!.appendChild(arraysNode!!.ownerDocument.createElement("string-array").also { arrayNode ->
-                arrayNode.setAttribute("name", arrayResource.name)
-
-                arrayResource.items.forEach { item ->
-                    item.include()
-
-                    arrayNode.appendChild(arrayNode.ownerDocument.createElement("item").also { itemNode ->
-                        itemNode.textContent = item.value
-                    })
-                }
-            })
-        }
-
-        /**
-         * Add a preference screen to the settings.
-         *
-         * @param preferenceScreen The name of the preference screen.
-         */
-        fun addPreferenceScreens(vararg preferenceScreen: PreferenceScreen) =
-            preferenceScreen.forEach { revancedPreferenceNode!!.addPreference(it) }
+        fun addArray(arrayResource: ArrayResource) =
+            arrayResource.include()
 
         /**
          * Add a preference fragment to the preferences.
@@ -111,105 +96,35 @@ class SettingsResourcePatch : ResourcePatch {
          * @param preference The preference to add.
          */
         fun addRootPreferences(vararg preference: BasePreference) =
-            preference.forEach { revancedPreferenceNode!!.addPreference(it) }
+            preference.forEach { revancedPreferenceNode!!.addPreference(it) { res -> res.include() } }
 
         /**
-         * Add a preference to the settings.
+         * Add a new resource to the resources.
          *
-         * @param preference The preference to add.
+         * @throws IllegalArgumentException if the resource already exists.
          */
-        private fun Node.addPreference(preference: BasePreference) {
-            // add a summary to the element
-            fun Element.addSummary(summaryResource: StringResource?, summaryType: SummaryType = SummaryType.DEFAULT) =
-                summaryResource?.let { summary ->
-                    setAttribute("android:${summaryType.type}", "@string/${summary.also { it.include() }.name}")
+        private fun IResource.include() {
+            when(this) {
+                is StringResource -> {
+                    if (strings.any { it.name == name }) return
+                    strings.add(this)
                 }
-
-            fun <T> Element.addDefault(default: T) {
-                default?.let {
-                    setAttribute(
-                        "android:defaultValue", when (it) {
-                            is Boolean -> if (it) "true" else "false"
-                            is String -> it
-                            else -> throw IllegalArgumentException("Unsupported default value type: ${it::class.java.name}")
-                        }
-                    )
+                is ArrayResource -> {
+                    if (arrays.any { it.name == name }) return
+                    arrays.add(this)
                 }
+                else -> throw NotImplementedError("Unsupported resource type")
             }
-
-            val preferenceElement = createElement(preference.tag)
-            preferenceElement.setAttribute("android:key", preference.key)
-            preferenceElement.setAttribute("android:title", "@string/${preference.title.also { it.include() }.name}")
-
-            when (preference) {
-                is PreferenceScreen -> {
-                    for (childPreference in preference.preferences) preferenceElement.addPreference(childPreference)
-                    preferenceElement.addSummary(preference.summary)
-                }
-                is PreferenceCategory -> {
-                    for (childPreference in preference.preferences) preferenceElement.addPreference(childPreference)
-                }
-                is SwitchPreference -> {
-                    preferenceElement.addDefault(preference.default)
-                    preferenceElement.addSummary(preference.summaryOn, SummaryType.ON)
-                    preferenceElement.addSummary(preference.summaryOff, SummaryType.OFF)
-                }
-                is TextPreference -> {
-                    preferenceElement.setAttribute("android:inputType", preference.inputType.type)
-                    preferenceElement.addDefault(preference.default)
-                    preferenceElement.addSummary(preference.summary)
-                }
-                is ListPreference -> {
-                    preferenceElement.setAttribute("android:entries", "@array/${preference.entries.also { addArray(it) }.name}")
-                    preferenceElement.setAttribute("android:entryValues", "@array/${preference.entryValues.also { addArray(it) }.name}")
-                    preferenceElement.addDefault(preference.default)
-                }
-            }
-
-            appendChild(preferenceElement)
-        }
-
-        /**
-         * Add a new string to the resources.
-         *
-         * @throws IllegalArgumentException if the string already exists.
-         */
-        private fun StringResource.include() {
-            if (strings.any { it.name == name }) return
-            strings.add(this)
-        }
-
-        /**
-         * Add a new array to the resources.
-         *
-         * @throws IllegalArgumentException if the string already exists.
-         */
-        private fun ArrayResource.include() {
-            if (arrays.any { it.name == name }) return
-            arrays.add(this)
         }
 
         private fun DomFileEditor?.getNode(tagName: String) = this!!.file.getElementsByTagName(tagName).item(0)
-
-        private fun Node?.createElement(tagName: String) = this!!.ownerDocument.createElement(tagName)
-
-        private enum class SummaryType(val type: String) {
-            DEFAULT("summary"), ON("summaryOn"), OFF("summaryOff")
-        }
     }
 
     override fun close() {
-        // merge all strings, skip duplicates
-        strings.forEach { stringResource ->
-            stringsNode!!.appendChild(stringsNode!!.ownerDocument.createElement("string").also { stringElement ->
-                stringElement.setAttribute("name", stringResource.name)
-
-                // if the string is un-formatted, explicitly add the formatted attribute
-                if (!stringResource.formatted) stringElement.setAttribute("formatted", "false")
-
-                stringElement.textContent = stringResource.value
-            })
+        arrays.forEach { array ->
+            arraysNode!!.addResource(array) { it.include() }
         }
+        strings.forEach { stringsNode!!.addResource(it) }
 
         revancedPreferencesEditor?.close()
         stringsEditor?.close()
